% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_spsimrec.R
\name{bym2_sp_rnd_ef}
\alias{bym2_sp_rnd_ef}
\title{bym2_sp_rnd_ef}
\usage{
bym2_sp_rnd_ef(nb_mat, sp_tau, sp_phi)
}
\arguments{
\item{nb_mat}{Matriz de vizinhança.}

\item{sp_tau}{Marginal precision of the spatial effect}

\item{sp_phi}{Proportion of spatial variance (structured vs unstructured)}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Gera efeitos aleatórios espaciais utilizando a estrutura do modelo BYM2.
}\if{html}{\out{</div>}}
}
