% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_spsimrec.R
\name{CAR.simWmat}
\alias{CAR.simWmat}
\title{CAR.simWmat}
\usage{
CAR.simWmat(sp_tau, sp_alpha, nb_mat)
}
\arguments{
\item{sp_tau}{Valor do parâmetro de precisão no modelo ICAR (\tau).}

\item{sp_alpha}{Valor do associação espacial no modelo CAR.}

\item{nb_mat}{Matriz de vizinhança.}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Gera efeitos aleatórios espaciais utilizando a decomposição de Cholesky.
}\if{html}{\out{</div>}}
}
