% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_spsimrec.R, R/Utils_spsimrec2.R
\name{car_sp_rnd_ef}
\alias{car_sp_rnd_ef}
\title{car_sp_rnd_ef}
\usage{
car_sp_rnd_ef(SP_N, sp_tau, sp_alpha, nb_mat)

car_sp_rnd_ef(SP_N, sp_tau, sp_alpha, nb_mat)
}
\arguments{
\item{SP_N}{Número de unidades de área.}

\item{sp_tau}{Valor do parâmetro de precisão no modelo ICAR (\tau).}

\item{sp_alpha}{Valor do associação espacial no modelo CAR.}

\item{nb_mat}{Matriz de vizinhança.}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Gera efeitos aleatórios espaciais considerando a estrutura do modelo CAR.
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{ Gera efeitos aleatórios espaciais considerando a estrutura do modelo CAR.
}\if{html}{\out{</div>}}
}
