% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sim.R
\name{func_sim}
\alias{func_sim}
\title{func_sim}
\usage{
func_sim(
  N,
  dist.x,
  par.x,
  beta.x,
  dist.z,
  par.z,
  dist.rec,
  par.rec,
  fun.min,
  fun.max,
  cens.prob,
  dfree,
  pfree,
  pfree1,
  logist,
  ent.dis.z,
  mu.z,
  sigma.z
)
}
\arguments{
\item{sigma.z}{desvio da normal para os efeitos aleatórios (caso ent.dis.z=1)}

\item{time:}{tempo até ocorrência}

\item{event:}{identificador de ocorrência de evento ou censura}

\item{id:}{identificador do indivíduo}

\item{IndRec:}{Indicador se o indivíduo apresenta uma ou mais recorrência}

\item{N:}{N}

\item{dist.x:}{Distribuição das covariáveis}

\item{par.x:}{Parâmetros das distribuições associadas às covariáveis}

\item{beta.x:}{Coeficientes associados às covariáveis}

\item{dist.z:}{Distribuição dos efeitos aleatórios}

\item{par.z:}{Parâmetros da distribuição dos efeitos aleatórios}

\item{dist.rec:}{Distribuição dos eventos recorrentes}

\item{par.rec:}{Parâmetros da distribuição dos eventos recorrentes}

\item{fun.min:}{Período de acompanhamento mínimo}

\item{fun.max:}{Período de acompanhamento máximo}

\item{cens.prob:}{Percentual de indivíduos censurados antes do fim do período de acompanhamento}

\item{dfree:}{Tempo fora de risco após experimentar um primeiro evento}

\item{pfree:}{Proporção de indivíduos fora de risco (sem utilizar covariáveis)}

\item{pfree1:}{Probabilidade de um indivíduo não recorrer (regressão logística)}

\item{logist:}{Uso da regressão logística}

\item{ent.dis.z:}{indicadora - entrada dos parâmetros da normal (1) ou da lognormal(0) para os efeitos aleatórios}

\item{mu.z:}{média da normal para os efeitos aleatórios (caso ent.dis.z=1)}
}
\description{
func_sim: Executa simulações
}
