% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_spsimrec.R, R/Utils_spsimrec2.R
\name{gen_data}
\alias{gen_data}
\title{gen_data}
\usage{
gen_data(
  ID,
  N,
  dist_int_func,
  par_int_func,
  fu,
  fu_max,
  x,
  beta_x_rec,
  tp_rnd_ef,
  rnd_ef,
  recurr,
  nr.cov_rec
)

gen_data(
  ID,
  N,
  dist_int_func,
  par_int_func,
  fu,
  fu_max,
  x,
  beta_x_rec,
  tp_rnd_ef,
  rnd_ef,
  recurr,
  nr.cov_rec
)
}
\arguments{
\item{ID}{Identificador dos indivíduos.}

\item{N}{Número de indivíduos.}

\item{dist_int_func}{Forma da função de intensidade.  "Weibull" (Lei de potência)}

\item{par_int_func}{Parâmetros da função de intensidade. Escala e forma (Lei de potência)}

\item{x}{Matriz de covariávies}

\item{rnd_ef}{Vetor de efeitos aleatórios.}

\item{recurr}{Indicador de recorrência.}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      Gera os tempos de recorrências dos eventos.
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{      Gera os tempos de recorrências dos eventos.
}\if{html}{\out{</div>}}
}
