% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_spsimrec.R
\name{gen_data_pol_int}
\alias{gen_data_pol_int}
\alias{gen_data_plp1}
\title{gen_data_plp1}
\usage{
gen_data_pol_int(ID, N, fu_max, x_cov, beta_x_rec, rnd_ef_tot, recurr)
}
\arguments{
\item{ID}{Identificador dos indivíduos.}

\item{N}{Número de indivíduos.}

\item{fu_max}{tempo máximo de acompanhamento.}

\item{x_cov}{Matrix de covariáveis.}

\item{rnd_ef_tot}{Vetor de efeitos aleatórios.}

\item{recurr}{Indicador de recorrência.}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      Gera os tempos de ocorrência e recorrência de todos os eventos.
}\if{html}{\out{</div>}}
}
