% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_spsimrec.R, R/Utils_spsimrec2.R
\name{gen_rnd_ef}
\alias{gen_rnd_ef}
\title{gen_rnd_ef_}
\usage{
gen_rnd_ef(N, ID, dist_z, tp_rnd_ef, par_z, mu_omega, sigma_omega)

gen_rnd_ef(N, ID, dist_z, tp_rnd_ef, par_z, mu_omega, sigma_omega)
}
\arguments{
\item{N}{Número de indivíduos.}

\item{ID}{Identificador de indivíduos.}

\item{dist_z}{Distribuição do efeito aleatório não espacial (Gamma ou lognormal).}

\item{tp_rnd_ef}{Indicadora de efeito aleatório multiplicativo.}

\item{par_z}{Parâmetros de variância da distribuição do efeito aleatório.}

\item{mu_omega}{Parâmetro desvio padrão da distribuição dos efeitos aleatórios para a forma {Y_i(t)*lambda_0(t)*exp(beta^t X_i+omega_i)}}

\item{random_ef}{Indicadora de uso de efeitos aleatórios na geração dos dados (random_ef==0, modelo sem efeitos).}
}
\description{
Gera efeitos aleatórios não espacial.

Gera efeitos aleatórios não espacial.
}
