% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_spsimrec.R, R/Utils_spsimrec2.R
\name{gen_zi}
\alias{gen_zi}
\alias{spsimrec}
\title{spsimrec}
\usage{
gen_zi(ID, N, pi, logist, cov_log, beta_x_log, xi1 = 1, rnd_ef)

gen_zi(ID, N, pi, logist, cov_log, beta_x_log, xi1 = 1, rnd_ef)
}
\arguments{
\item{N}{Número de indivíduos.}

\item{pi}{Inflação de zeros (proporção de indivíduos sem recorrências).}

\item{cov_log}{Lista de covariáveis relacionadas à regressão logística.}

\item{beta_x_log}{Coeficientes das covariáveis relacionadas à regressão logística.}

\item{xi}{Parâmetro que relaciona a função de intensidade e a logística.}

\item{rnd}{Vetor de efeitos aleatórios.}
}
\description{
Recur:

Recur:
}
