% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cov_fu1.R
\name{gencovfu1}
\alias{gencovfu1}
\alias{gencovfu}
\title{gencovfu}
\usage{
gencovfu1(
  N,
  fu.min,
  fu.max,
  cens.prob = 0,
  dist.x = NULL,
  par.x = NULL,
  beta.x = NULL
)
}
\arguments{
\item{cens.prob}{Probabiidade de apresentar censura antes do período máximo de acompanhamento.}

\item{dist.x}{Distribuição das Covariáveis. Binomial ou normal.}

\item{par.x}{Parâmetros das distribuições das covariáveis.}

\item{beta.x}{Coeficiente de regressão das covariáveis.}
}
\description{
Recur:
}
