% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_spsimrec.R
\name{get_nhpp_realization}
\alias{get_nhpp_realization}
\title{get_nhpp_realization}
\usage{
get_nhpp_realization(lambda)
}
\arguments{
\item{lambda}{Funćão de intensidade polinomial.}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      Gera os tempos de ocorrência das reincidências.
}\if{html}{\out{</div>}}
}
