% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_spsimrec.R
\name{get_nhpp_realization_0}
\alias{get_nhpp_realization_0}
\title{get_nhpp_realization_0}
\usage{
get_nhpp_realization_0(lambda, fu_max)
}
\arguments{
\item{lambda}{Funćão de intensidade polinomial.}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      Gera os tempos de ocorrência dos primeiros eventos.
}\if{html}{\out{</div>}}
}
