% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_spsimrec.R
\name{icar_sp_rnd_ef}
\alias{icar_sp_rnd_ef}
\title{icar_sp_rnd_ef}
\usage{
icar_sp_rnd_ef(nb_mat, sig = 1)
}
\arguments{
\item{nb_mat}{Matriz de vizinhança.}

\item{sig}{Parâmetro desvio padrão dos efeitos aleatórios.}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Gera efeitos aleatórios espaciais utilizando a estrutura do modelo ICAR.
}\if{html}{\out{</div>}}
}
