% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recur.R
\name{Recur}
\alias{Recur}
\title{Recur}
\usage{
Recur(time, event = NULL, id = NULL, IndRec = NULL)
}
\arguments{
\item{time:}{tempo até ocorrência}

\item{event:}{identificador de ocorrência de evento ou censura}

\item{id:}{identificador do indivíduo}

\item{IndRec:}{Indicador se o indivíduo apresenta uma ou mais recorrência}
}
\description{
Recur: Preparação dos dados
}
