% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spnhppzi-package.R
\name{spnhppzi-package}
\alias{spnhppzi-package}
\alias{spnhppzi}
\title{The 'spnhppzi' package.}
\description{
A DESCRIPTION OF THE PACKAGE

_PACKAGE
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.5. https://mc-stan.org
}
