% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spnhppzi.R
\name{spnhppzi}
\alias{spnhppzi}
\title{spnhppzi}
\usage{
spnhppzi(
  formula,
  data,
  baseline = c("plp1", "plp2"),
  approach = c("mle", "bayes"),
  n_iter = 4000,
  ZI = c("true", "false"),
  FR = c("true", "false"),
  initial,
  tp_prior = 0,
  frag = 0,
  mu_omega = 0,
  sigma_omega = 0,
  shp_sigma_omega = 0,
  scl_sigma_omega = 0,
  shp_alpha1 = 0,
  scl_alpha1 = 0,
  shp_alpha2 = 0,
  scl_alpha2 = 0,
  mu_beta = 0,
  sigma_beta = 10,
  mu_psi = 0,
  sigma_psi = 10
)
}
\arguments{
\item{data:}{data frame contendo conjunto de dados}

\item{id:}{identificador de indivíduo}

\item{status:}{evento ou censura}

\item{stop:}{tempo até ocorrência do evento}

\item{IndRec:}{Indicador se o indivíduo apresenta uma ou mais recorrência}

\item{covar:}{Covariável}

\item{initial:}{Valores de incialização dos parâmetros}

\item{approach:}{abordagem a ser considerada (0 - frequentista, 1 - bayesiana)}

\item{frag:}{uso de fragilidade (0 - sem fragilidade, 1 - com fragilidade)}
}
\description{
NHPPZI: Ajusta um modelo de processo de poisson não homegêneo
}
