% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spnhppzi4.R, R/spnhppzi5.R
\name{spnhppzi4}
\alias{spnhppzi4}
\alias{spnhppzi}
\title{spnhppzi}
\usage{
spnhppzi4(
  formula,
  data,
  baseline = c("plp1", "plp2", "plp3", "bp"),
  approach = c("mle", "bayes"),
  n_iter = 4000,
  n_cores = 1,
  n_chains = 4,
  ZI = c("true", "false"),
  FR = c("true", "false"),
  initial,
  tp_prior = 0,
  frag = 0,
  tp_rnd_ef = 0,
  mu_omega = 0,
  sigma_omega = 0,
  shp_sigma2_z = 0,
  scl_sigma2_z = 0,
  shp_alpha1 = 0,
  scl_alpha1 = 0,
  shp_alpha2 = 0,
  scl_alpha2 = 0,
  mu_beta = 0,
  sigma_beta = 10,
  mu_psi = 0,
  sigma_psi = 10,
  rnd_logist = 0,
  mu_xi = 0,
  sigma_xi = 1,
  spatial = 0,
  sp_model = c("car", "sparse", "icar"),
  nb_mat = NULL,
  data_tau = 0,
  tau = 1,
  tp_prior_tau = c("gamma", "inv_gamma"),
  shp_tau = 0,
  scl_tau = 0,
  lower_tau = 0,
  tp_icar = 0,
  W_n = 0,
  bp_degree = NULL,
  h1_gamma = 0,
  h2_gamma = 4,
  omega_data = 0,
  omega = NULL,
  std_dev = 1,
  tp_DIC = 0
)

spnhppzi4(
  formula,
  data,
  baseline = c("plp1", "plp2", "plp3", "bp"),
  approach = c("mle", "bayes"),
  n_iter = 4000,
  n_cores = 1,
  n_chains = 4,
  ZI = c("true", "false"),
  FR = c("true", "false"),
  initial,
  tp_prior = 0,
  frag = 0,
  tp_rnd_ef = 0,
  mu_omega = 0,
  sigma_omega = 0,
  shp_sigma2_z = 0,
  scl_sigma2_z = 0,
  shp_alpha1 = 0,
  scl_alpha1 = 0,
  shp_alpha2 = 0,
  scl_alpha2 = 0,
  mu_beta = 0,
  sigma_beta = 10,
  mu_psi = 0,
  sigma_psi = 10,
  rnd_logist = 0,
  mu_xi = 0,
  sigma_xi = 1,
  spatial = 0,
  sp_model = c("car", "sparse", "icar"),
  nb_mat = NULL,
  data_tau = 0,
  tau = 1,
  tp_prior_tau = c("gamma", "inv_gamma"),
  shp_tau = 0,
  scl_tau = 0,
  lower_tau = 0,
  tp_icar = 0,
  W_n = 0,
  bp_degree = NULL,
  h1_gamma = 0,
  h2_gamma = 4,
  omega_data = 0,
  omega = NULL,
  std_dev = 1,
  tp_DIC = 0
)
}
\arguments{
\item{data:}{data frame contendo conjunto de dados}

\item{id:}{identificador de indivíduo}

\item{status:}{evento ou censura}

\item{stop:}{tempo até ocorrência do evento}

\item{IndRec:}{Indicador se o indivíduo apresenta uma ou mais recorrência}

\item{covar:}{Covariável}

\item{initial:}{Valores de incialização dos parâmetros}

\item{approach:}{abordagem a ser considerada (0 - frequentista, 1 - bayesiana)}

\item{frag:}{uso de fragilidade (0 - sem fragilidade, 1 - com fragilidade)}
}
\description{
NHPPZI: Ajusta um modelo de processo de poisson não homegêneo

NHPPZI: Ajusta um modelo de processo de poisson não homegêneo
}
