% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsimrec.R
\name{spsimrec}
\alias{spsimrec}
\title{spsimrec}
\usage{
spsimrec(
  N,
  fu.min,
  fu.max,
  cens.prob = 0,
  dist.x = NULL,
  par.x = NULL,
  beta.x = NULL,
  dist.z = c("gamma", "lognormal"),
  random.ef = 0,
  tp_rnd_ef = 0,
  par.z = 0,
  dist.rec,
  par.rec = 0,
  pi = 0,
  dfree = 0,
  logist = 0,
  mu.omega = 0,
  sigma.omega = 0
)
}
\arguments{
\item{cens.prob}{Probabiidade de apresentar censura antes do período máximo de acompanhamento.}

\item{dist.x}{Distribuição das Covariáveis. Binomial ou normal.}

\item{par.x}{Parâmetros das distribuições das covariáveis.}

\item{beta.x}{Coeficiente de regressão das covariáveis.}

\item{dist.z}{Distribuição do efeito aleatório. Gamma ou lognormal.}

\item{par.z}{Parâmetros da distribuição do efeito aleatório.}

\item{dist.rec}{Forma da função de intensidade.  "Weibull" (Lei de potência)}

\item{par.rec}{Parâmetros da função de intensidade. Escala e forma}
}
\description{
Recur:
}
