% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsimrec1.R
\name{spsimrec1}
\alias{spsimrec1}
\alias{spsimrec}
\title{spsimrec}
\usage{
spsimrec1(
  N,
  x,
  x1,
  cov.rec,
  cov.log,
  beta.x.rec = NULL,
  beta.x.log = NULL,
  fu,
  fu.max,
  fu.min,
  dist.z = c("gamma", "lognormal"),
  random.ef = 0,
  tp_rnd_ef = 0,
  par.z = 0,
  dist.rec,
  par.rec = 0,
  pi = 0,
  dfree = 0,
  logist = 0,
  mu.omega = 0,
  sigma.omega = 0
)
}
\arguments{
\item{dist.z}{Distribuição do efeito aleatório. Gamma ou lognormal.}

\item{par.z}{Parâmetros da distribuição do efeito aleatório.}

\item{dist.rec}{Forma da função de intensidade.  "Weibull" (Lei de potência)}

\item{par.rec}{Parâmetros da função de intensidade. Escala e forma}

\item{beta.x}{Coeficiente de regressão das covariáveis.}
}
\description{
Recur:
}
