% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsimrec2.R
\name{spsimrec2}
\alias{spsimrec2}
\alias{spsimrec}
\title{spsimrec}
\usage{
spsimrec2(
  N,
  nr.cov,
  spatial,
  sp_model = c("car", "icar"),
  list_area,
  SP_N,
  nb_mat,
  sp_tau,
  sp_alpha,
  beta.x = NULL,
  x,
  x1,
  fu,
  fu.max,
  fu.min,
  dist.z = c("gamma", "lognormal"),
  random.ef = 0,
  tp_rnd_ef = 0,
  par.z = 0,
  dist.rec,
  par.rec = 0,
  pi = 0,
  dfree = 0,
  logist = 0,
  mu.omega = 0,
  sigma.omega = 0
)
}
\arguments{
\item{beta.x}{Coeficiente de regressão das covariáveis.}

\item{dist.z}{Distribuição do efeito aleatório. Gamma ou lognormal.}

\item{par.z}{Parâmetros da distribuição do efeito aleatório.}

\item{dist.rec}{Forma da função de intensidade.  "Weibull" (Lei de potência)}

\item{par.rec}{Parâmetros da função de intensidade. Escala e forma}
}
\description{
Recur:
}
