% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsimrec8.1.R
\name{spsimrec8.1}
\alias{spsimrec8.1}
\alias{spsimrec}
\title{spsimrec}
\usage{
spsimrec8.1(
  N,
  spatial = 0,
  sp_model = c("car", "sparse", "icar"),
  list_area = NULL,
  SP_N = NULL,
  nb_mat = NULL,
  sp_tau = NULL,
  sp_alpha = NULL,
  x1,
  cov_rec = NULL,
  cov_log = NULL,
  beta_x_rec = NULL,
  beta_x_log = NULL,
  xi = 1,
  fu,
  fu_max,
  fu_min,
  random_ef = 0,
  dist_z = c("gamma", "lognormal"),
  tp_rnd_ef = 0,
  par_z = 0,
  dist_int_func,
  par_int_func = 0,
  pi = 0,
  logist = 0,
  mu_omega = 0,
  sigma_omega = 0,
  baseline = c("plp1", "plp2")
)
}
\arguments{
\item{N}{Número de indivíduos}

\item{spatial}{Indicadora do uso de modelo espacial, (default =0 , não espacial)}

\item{sp_model}{Tipo de modelo espacial}

\item{SP_N}{Número de unidades de área}

\item{nb_mat}{Matriz de vizinhança}

\item{sp_tau}{Valor do parâmetro de precisão no modelo ICAR (\tau)}

\item{sp_alpha}{Valor do associação espacial no modelo CAR.}

\item{x1}{Matriz ID e covariávies}

\item{cov_rec}{Lista de covariáveis relacionadas à função de intensidade}

\item{cov_log}{Lista de covariáveis relacionadas à regressão logística}

\item{beta_x_rec}{Coeficientes das covariáveis relacionadas à função de intensidade}

\item{beta_x_log}{Coeficientes das covariáveis relacionadas à regressão logística}

\item{xi}{Parâmetro que relaciona a função de intensidade e a logística}

\item{fu_max}{Tempo máximo de acompanhamento}

\item{fu_min}{Tempo mínimo de acompanhamento}

\item{random_ef}{Indicadora de uso de efeitos aleatórios na geração dos dados (random_ef==0, modelo sem efeitos).}

\item{dist_z}{Distribuição do efeito aleatório não espacial (Gamma ou lognormal).}

\item{tp_rnd_ef}{Indicadora de efeito aleatório multiplicativo}

\item{par_z}{Parâmetros de variância da distribuição do efeito aleatório.}

\item{dist_int_func}{Forma da função de intensidade.  "Weibull" (Lei de potência)}

\item{par_int_func}{Parâmetros da função de intensidade. Escala e forma (Lei de potência)}

\item{pi}{Inflação de zeros (proporção de indivíduos sem recorrências)}

\item{logist}{Indicadora do uso de covariáveis na logística}

\item{x}{Matriz de covariávies}
}
\description{
Recur:
}
