% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recur.R
\name{Recur}
\alias{Recur}
\title{Recur}
\usage{
Recur(time, event = NULL, id = NULL, SP_ID = NULL, IndRec = NULL)
}
\arguments{
\item{time}{Numeric. Time until the occurrence of an event.}

\item{event}{Integer (0 or 1). Indicator of event occurrence (\code{1}) or censoring (\code{0}).}

\item{id}{Integer or Character. Unique identifier for each individual.}

\item{IndRec}{Integer (0 or 1). Indicator of whether the individual has at least one recurrence (\code{1}) or not (\code{0}).}
}
\description{
Prepares recurrent event data for analysis by structuring key variables,
including event times, censoring indicators, and recurrence status.
}
