% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsimrec.R
\name{spsimrec}
\alias{spsimrec}
\title{spsimrec}
\usage{
spsimrec(
  N,
  spatial = 0,
  sp_model = c("car", "sparse", "icar"),
  list_area = NULL,
  SP_N = NULL,
  nb_mat = NULL,
  sp_tau = NULL,
  sp_alpha = NULL,
  x1,
  cov_rec = NULL,
  cov_log = NULL,
  beta_x_rec = NULL,
  beta_x_log = NULL,
  xi = 1,
  fu,
  fu_max,
  fu_min,
  random_ef = 0,
  dist_z = c("gamma", "lognormal"),
  tp_rnd_ef = 0,
  par_z = 0,
  dist_int_func,
  par_int_func = 0,
  pi_zi = 0,
  logist = 0,
  mu_omega = 0,
  sigma_omega = 0,
  baseline = c("plp", "polynom")
)
}
\arguments{
\item{N}{Integer. Number of individuals in the dataset.}

\item{spatial}{Logical (0 or 1). Indicator for spatial modeling (default = 0, non-spatial).}

\item{sp_model}{Character. Type of spatial model used (\code{"ICAR"} or \code{"CAR"}).}

\item{SP_N}{Integer. Number of area units when using a spatial model.}

\item{nb_mat}{Matrix. Neighborhood matrix defining spatial relationships.}

\item{sp_tau}{Numeric. Precision parameter (\code{τ}) for the ICAR model.}

\item{sp_alpha}{Numeric. Spatial association parameter for the CAR model.
\subsection{Covariates}{
}}

\item{x1}{Matrix. Matrix containing \code{ID} and covariates.}

\item{cov_rec}{Character vector. List of covariates associated with the intensity function.}

\item{cov_log}{Character vector. List of covariates associated with the logistic regression.}

\item{beta_x_rec}{Numeric vector. Coefficients of covariates in the intensity function.}

\item{beta_x_log}{Numeric vector. Coefficients of covariates in the logistic regression.
\subsection{Follow-up and Zero Inflation}{
}}

\item{xi}{Numeric. Parameter linking the intensity function and logistic regression.
\subsection{Logistic Regression Model}{
}}

\item{fu_max}{Numeric. Maximum follow-up time.}

\item{fu_min}{Numeric. Minimum follow-up time.}

\item{random_ef}{Logical (0 or 1). Indicator for using random effects (default = 0, no random effects).}

\item{dist_z}{Character. Distribution of the non-spatial random effect (\code{"Gamma"} or \code{"lognormal"}).}

\item{tp_rnd_ef}{Logical (0 or 1). Indicator for using multiplicative random effects.}

\item{par_z}{Numeric vector. Variance parameters for the random effect distribution.
\subsection{Intensity Function}{
}}

\item{dist_int_func}{Character. Type of intensity function (\code{"Weibull"} for power-law process).}

\item{par_int_func}{Numeric vector. Scale and shape parameters of the intensity function.}

\item{pi_zi}{Numeric. Proportion of individuals with zero recurrences (zero inflation).}

\item{logist}{Logical (0 or 1). Indicator for using covariates in the logistic regression model.}

\item{x}{Matrix. Matrix of covariates.}

\item{cens.prob}{Numeric. Probability of censoring before the maximum follow-up period.
\subsection{Random Effects}{
}}

\item{dist.x}{Character. Distribution of covariates (\code{"binomial"} or \code{"normal"}).}

\item{par.x}{List. Parameters of the covariate distributions.}
}
\description{
Simulates recurrent event data with options for spatial correlation, zero inflation,
random effects, and different intensity functions. The function allows customization
of the follow-up period, covariates, and model parameters.
}
